Move common utils to indexer_util
