+++ bigquery/requirements-to-freeze.txt
@@ -1,3 +1,4 @@
+../indexer_util
 elasticsearch==6.1.1
 jsmin==2.2.2
 pandas==0.22.0
