+++ bigquery/requirements.txt
@@ -1,28 +1,29 @@
-cachetools==2.0.1
-certifi==2018.1.18
+cachetools==2.1.0
+certifi==2018.4.16
 chardet==3.0.4
 elasticsearch==6.1.1
 futures==3.2.0
-google-api-core==1.0.0
-googleapis-common-protos==1.5.3
-google-auth==1.4.1
+google-api-core==1.2.1
+google-auth==1.5.0
 google-auth-oauthlib==0.2.0
-google-cloud-bigquery==0.31.0
+google-cloud-bigquery==1.3.0
 google-cloud-core==0.28.1
 google-resumable-media==0.3.1
-idna==2.6
+googleapis-common-protos==1.5.3
+idna==2.7
+./indexer_util
 jsmin==2.2.2
-numpy==1.14.1
-oauthlib==2.0.6
+numpy==1.14.5
+oauthlib==2.1.0
 pandas==0.22.0
 pandas-gbq==0.3.1
-protobuf==3.5.2
-pyasn1==0.4.2
+protobuf==3.6.0
+pyasn1==0.4.3
 pyasn1-modules==0.2.1
-python-dateutil==2.6.1
-pytz==2018.3
-requests==2.18.4
-requests-oauthlib==0.8.0
+python-dateutil==2.7.3
+pytz==2018.4
+requests==2.19.1
+requests-oauthlib==1.0.0
 rsa==3.4.2
 six==1.11.0
 urllib3==1.22
