+++ indexer_util/README.md
@@ -1,0 +1,43 @@
+# Indexer Utilities
+
+Utilities that can be used by any indexer.
+
+## Installation
+
+Add to indexer's `requirements-to-freeze.txt`:
+```
+../indexer_util
+```
+
+By default, `pip freeze` will convert this line to `indexer_util=x.y.z` in
+`requirements.txt`, but this causes problems when trying to `pip install` the
+requirements. So you should convert it back to ./indexer_util.
+You can do this using `sed` in conjunction with `pip freeze`:
+```
+pip freeze | sed 's/^indexer_util.*/\.\/indexer_util/g' > requirements.txt
+```
+
+Add to indexer's Dockerfile:
+```
+ADD indexer_util /app/indexer_util
+```
+
+The indexer can import indexer_util like this:
+```
+import indexer_util
+```
+
+## Generating `requirements.txt`
+
+`requirements.txt` is autogenerated from `requirements-to-freeze.txt`. The
+latter lists only direct dependencies. To regenerate, run:
+```
+virtualenv --python=/usr/bin/python2 ~/virtualenv/indexer-util
+source ~/virtualenv/indexer-util/bin/activate
+pip install -r requirements-to-freeze.txt
+pip freeze > requirements.txt
+deactivate
+```
+
+Any library dependencies also need to be added to the `REQUIRES` list in
+`setup.py`
