+++ indexer_util/requirements-to-freeze.txt
@@ -1,0 +1,2 @@
+elasticsearch==6.1.1
+jsmin==2.2.2
