+++ indexer_util/requirements.txt
@@ -1,0 +1,3 @@
+elasticsearch==6.1.1
+jsmin==2.2.2
+urllib3==1.22
